Added new schema range attributes and updated related classes
A series of updates were made to improve the handling of schema ranges. New attributes `SchemaLengthRangeAttribute`, `SchemaPropertiesRangeAttribute`, and `SchemaValueRangeAttribute` were added to specify length, property, and value ranges for a schema. The `SchemaMaxAttribute` and `SchemaMinAttribute` classes were removed and replaced with these new attributes in `JsonSchemaBuilderExtensions.cs` and `FallbackTypeHandler.cs`.

A new extension method `TryGetCustomAttributeData<T>` was added to `MemberInfoExtensions.cs` and the method `GetPropertyName` in `PropertyInfoExtensions.cs` was updated. A new method `FromType<T>` was added to `RootTypeContext.cs`.

A new class `StringExtensions.cs` was added with an extension method `ToJsonPropertyName` for camel case formatting. The constructor in `TypeConverter.cs` was updated to accept null options and use default options if null is passed.

Unit tests were added in a new class `SchemaAttributeTests` to verify the functionality of the new schema attributes. A new base class `TestBase` was added for common test functionality.