Updated project structure, READMEs, and added new features to SharpSchema library
Significant updates were made to the SharpSchema library. The `Directory.Build.props` file was updated to point to the new repository URL. The `README.md` files were updated and created for various components to provide detailed information about the library and its usage. The project type of `SharpSchema.Tool` and `SharpSchema.Annotations` was changed and a new project `SharpSchema.Annotations.Source` was added. New properties were added to the `.csproj` files of `SharpSchema.Annotations` and `SharpSchema.Tool`. The visibility of classes in `SharpSchema.Annotations` was conditionally changed.  The `SharpSchema.csproj` file was updated to make the project packable and add package metadata.