Add ExcludeFromCodeCoverage attributes and JsonIgnore
This commit introduces the `[ExcludeFromCodeCoverage]` attribute
to several classes across multiple files to improve test coverage
accuracy. The attribute is added to `UnsupportedObjectKeyword`,
`JsonSchemaExtensions`, and `Throw` classes.

Additionally, the `AlsoIgnored` property in `Class_WithIgnoredProperty`
is marked with `[JsonIgnore(Condition = JsonIgnoreCondition.Always)]`
to ensure it is ignored during JSON serialization.

Finally, a reference to the `JsonIgnoreAttribute` assembly is added
in `TestDataFixture` to support the new JSON ignore functionality.